requirements.txt
flask
pymongo
python-dotenv
scapy
echarts-python